moved together test for dump and restore GRO, Ref. #113
