+++ mrmd/io/CMakeLists.txt
@@ -15,11 +15,8 @@
 
 mrmd_add_test(NAME test.io.DumpCSV
               FILES DumpCSV.test.cpp)
-configure_file(DumpGRO.test.gro ${CMAKE_CURRENT_BINARY_DIR} COPYONLY)
-mrmd_add_test(NAME test.io.DumpGRO
-              FILES DumpGRO.test.cpp)
-mrmd_add_test(NAME test.io.RestoreGRO
-              FILES RestoreGRO.test.cpp)
+mrmd_add_test(NAME test.io.GRO
+              FILES GRO.test.cpp)
 mrmd_add_test(NAME test.io.ThermoForce
               FILES ThermoForce.test.cpp)
