--- mrmd/io/DumpGRO.test.gro
@@ -1,23 +1,0 @@
-Test, t=0
-20
-    1Atom    At    1   0.000   0.000   0.000  0.0000  0.0000  0.0000
-    2Atom    At    2   0.000   0.000   0.000  0.0000  0.0000  0.0000
-    3Atom    At    3   0.000   0.000   0.000  0.0000  0.0000  0.0000
-    4Atom    At    4   0.000   0.000   0.000  0.0000  0.0000  0.0000
-    5Atom    At    5   0.000   0.000   0.000  0.0000  0.0000  0.0000
-    6Atom    At    6   0.000   0.000   0.000  0.0000  0.0000  0.0000
-    7Atom    At    7   0.000   0.000   0.000  0.0000  0.0000  0.0000
-    8Atom    At    8   0.000   0.000   0.000  0.0000  0.0000  0.0000
-    9Atom    At    9   0.000   0.000   0.000  0.0000  0.0000  0.0000
-   10Atom    At   10   0.000   0.000   0.000  0.0000  0.0000  0.0000
-   11Atom    At   11   0.000   0.000   0.000  0.0000  0.0000  0.0000
-   12Atom    At   12   0.000   0.000   0.000  0.0000  0.0000  0.0000
-   13Atom    At   13   0.000   0.000   0.000  0.0000  0.0000  0.0000
-   14Atom    At   14   0.000   0.000   0.000  0.0000  0.0000  0.0000
-   15Atom    At   15   0.000   0.000   0.000  0.0000  0.0000  0.0000
-   16Atom    At   16   0.000   0.000   0.000  0.0000  0.0000  0.0000
-   17Atom    At   17   0.000   0.000   0.000  0.0000  0.0000  0.0000
-   18Atom    At   18   0.000   0.000   0.000  0.0000  0.0000  0.0000
-   19Atom    At   19   0.000   0.000   0.000  0.0000  0.0000  0.0000
-   20Atom    At   20   0.000   0.000   0.000  0.0000  0.0000  0.0000
-    nan nan nan
